scripts 添加mydan/perl/.lock的判断

+++ scripts/uninstall.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
+if [ -f $BP/perl/.lock ]; then
+    echo "The perl is locked"
+    exit;
+fi
+
 rm -rf /opt/mydan/perl
